Add modular structure to js
