+++ src/modules/module3.js
@@ -1,0 +1,13 @@
+// const tasksArray = [];
+
+// export default tasksArray;.
+
+class Tasks {
+  constructor(description, completed, index) {
+    this.description = description;
+    this.completed = completed;
+    this.index = index;
+  }
+}
+
+export default Tasks;
