+++ src/style.css
@@ -25,7 +25,7 @@ body {
   font-size: 1rem;
 }
 
-.big-container input {
+.enter {
   border: none;
   outline: none;
   font-style: italic;
@@ -77,4 +77,10 @@ span {
 
 .edit-input {
   background-color: #ffffbf;
+  padding: 0;
+  border: none;
+  outline: none;
+  font-style: italic;
+  font-weight: 300;
+  width: 100%;
 }
